Created a custom list_view.xml, which is the layout for a list element with 2 TextViews
Created a custom Object Word.java which holds the miwok and default language translation in 2 varaibles as a state and has get methods for both variables.

Created a WordAdapter.java class as custom Adapter which extends the ArrayAdapter class:
The constructor takes the context and a Word Object as input. It creates a listItemView based on the list_item.xml layout.
It populates the miwok_text_view and default_text_view with the respective values of the Word object.
It then returns the listItemView