Added a constant to Word.java, which is assigned to mImageResourceId to have a mark if there is no image assigned. Added a hasImage method which checks the state of mImageResourceId and returns true, if an image is assigned, false if not. Added an if statemnt in WordAdapter.java to check if there is a valid image in the Word objecgt. If true, it will set the image to the ImageView and set it to visible (if set to gone in a recycled view). If false, the ImageView is set to gone.
Added visual polishes: background colors to the list_item.xml, borders and postioning of the TextViews, sizes of the views and text style and color. Removed borders in the word_list.xml, as the list_items should fit the whole screen.

Specific background colors for the activities will be set in next step.