Added functionality to display different background colors on the textes of the list items:
Made the constructor of the WordAdapter.java Class to take in a color resource ID.
Changed the WordAdapter constructions in the respective Category Activities to pass in the color resource.
Named the LinearLayout which holds the TextViews in list_item.xml (text_container).
Added code in WordAdapter.java to find the container and then set the background color on that.